add_tags/pyproject.toml
[tool.poetry]
name = "add-tags"
version = "0.1.0"
description = "A project to add tags to data using OpenAI."
authors = ["bjpl <brandon.lambert87@gmail.com>"]
license = ""

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.0"
openai = "^0.28.1"
python-dotenv = "^1.0.0"
transformers = "^4.36.0"
accelerate = "^0.26.0"
torch = "^2.1.0"
torchvision = "^0.16.0"
torchaudio = "^2.1.0"
numpy = "<2.0"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Specify the package directory
packages = [{include = "add_tags"}]